[REFACTOR] memisahkan attribute tentang keluarga ke class FamilyInformation
